function.h/cpp: use C++11 instead of autogenerated code
While working on function.{hppy,cppy} I've found it a bit difficult
to modify them. The template engine (yaptu.py) supports a limited
subset of python syntax, and is not documented anywhere. The error
messages are not helpful either. First I decided to replace yaptu
with a well known (python) template engine - jinja2, mako, etc.
However I've end up using C++ variadic templates instead. To keep
the code C++11 compatible I make use of mapbox/variant library which
provides `std::variant` equivalent. For convenience that library
is included as a git submodule. It can be removed if/when we upgrade
to C++17

As a side effect python is no longer required to build the library
(but is still necessary for building ginsh).